refactor(runtimed-service): extract service module from runtimed-client
Cross-platform daemon installation glue (launchd / SMAppService /
systemd / Windows Startup) lives in its own crate now.

- runtimed-client drops smappservice-rs as a direct dep.
- runtimed (daemon) and runt (CLI) gain runtimed-service as a path
  dep. The daemon re-exports it as runtimed::service so its main.rs
  paths stay unchanged.
- runt switches its call sites from runtimed::service:: to
  runtimed_service::.
