+++ crates/runtimed-client/Cargo.toml
@@ -36,9 +36,6 @@ automerge-recovery = { path = "../automerge-recovery" }
 schemars = { workspace = true }
 ts-rs = { workspace = true, optional = true }
 
-[target.'cfg(target_os = "macos")'.dependencies]
-smappservice-rs = "0.1"
-
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
