+++ crates/runtimed/src/lib.rs
@@ -16,6 +16,7 @@
 #![cfg_attr(test, allow(clippy::unwrap_used, clippy::expect_used))]
 
 pub use runtimed_client::*;
+pub use runtimed_service as service;
 
 // ============================================================================
 // Server-only modules (not in runtimed-client)
